# Improve performances: Remove the `aggregateAsync` call