--- a/zio-kafka/src/main/scala/zio/kafka/consumer/internal/Runloop.scala
+++ b/zio-kafka/src/main/scala/zio/kafka/consumer/internal/Runloop.scala
@@ -491,11 +491,14 @@ private[consumer] final class Runloop private (
           s"Starting poll with ${state.pendingRequests.size} pending requests and ${state.pendingCommits.size} pending commits"
         )
 
-    ZStream
-      .fromQueue(commandQueue)
-      .timeoutFail[Throwable](RunloopTimeout)(runloopTimeout)
-      .takeWhile(_ != StopRunloop)
-      .runChunksFoldZIO(State.initial) { case (state, commands) =>
+    /**
+     * Inlined, simplified and specialized for our needs version of [[ZSink.foldChunksZIO]]
+     *
+     * Code initially inspired by the implementation of [[ZStream.runFoldZIO]] with everything we don't need removed and
+     * with chunking added
+     */
+    val sink = {
+      def execute(state: State, commands: Chunk[Command]): Task[State] =
         for {
           _          <- ZIO.logTrace(s"Processing ${commands.size} commands: ${commands.mkString(",")}")
           isShutdown <- isShutdown
@@ -508,31 +511,28 @@ private[consumer] final class Runloop private (
           // Immediately poll again, after processing all new queued commands
           _ <- commandQueue.offer(Command.Poll).when(updatedStateAfterPoll.shouldPoll)
         } yield updatedStateAfterPoll
-      }
+
+      def reader(s: State): ZChannel[Any, Throwable, Chunk[Command], Any, Throwable, Nothing, Unit] =
+        ZChannel.readWith(
+          (in: Chunk[Command]) => ZChannel.fromZIO(execute(s, in)).flatMap(reader),
+          (err: Throwable) => ZChannel.fail(err),
+          (_: Any) => ZChannel.unit
+        )
+
+      ZSink.fromChannel(reader(State.initial))
+    }
+
+    ZStream
+      .fromQueue(commandQueue)
+      .timeoutFail[Throwable](RunloopTimeout)(runloopTimeout)
+      .takeWhile(_ != StopRunloop)
+      .run(sink)
       .tapErrorCause(cause => ZIO.logErrorCause("Error in Runloop", cause))
       .onError(cause => partitions.offer(Take.failCause(cause)))
   }
 }
 
 private[consumer] object Runloop {
-  private implicit final class RichZStream[R, E, A](private val stream: ZStream[R, E, A]) extends AnyVal {
-
-    /**
-     * Adapted from [[ZStream.runFoldZIO]]
-     */
-    def runChunksFoldZIO[R1 <: R, E1 >: E, S](s: => S)(f: (S, Chunk[A]) => ZIO[R1, E1, S])(implicit
-      trace: Trace
-    ): ZIO[R1, E1, S] = ZIO.scoped[R1](runChunksFoldWhileScopedZIO[R1, E1, S](s)(_ => true)(f))
-
-    /**
-     * Adapted from [[ZStream.runFoldWhileScopedZIO]]
-     */
-    private def runChunksFoldWhileScopedZIO[R1 <: R, E1 >: E, S](
-      s: => S
-    )(cont: S => Boolean)(f: (S, Chunk[A]) => ZIO[R1, E1, S])(implicit trace: Trace): ZIO[R1 with Scope, E1, S] =
-      stream.runScoped(ZSink.foldChunksZIO(s)(cont)(f))
-  }
-
   type ByteArrayCommittableRecord = CommittableRecord[Array[Byte], Array[Byte]]
 
   // Internal parameters, should not be necessary to tune