@svroonland 2023-03-27
Benchmark results are within each others' confidence ranges, but keeping chunks seems like a good change anyway.